AutoSub/AutoSub/Resources/backend/requirements.txt
deepgram-sdk>=5.3.2
google-genai>=1.61.0
python-dotenv>=1.2.1
websockets>=13.0
